add README to dynamodb-standalone example
also:
- add some comment in code for better understanding
- add console output when running tests with gradle
- add application plugin for run with gradle
