+++ dynamodb/dynamodb-standalone/src/main/java/io/flamingock/examples/dynamodb/standalone/CommunityStandaloneDynamoDBApp.java
@@ -16,33 +16,39 @@
 
 package io.flamingock.examples.dynamodb.standalone;
 
-import io.flamingock.core.configurator.core.CoreConfiguration;
 import io.flamingock.core.configurator.standalone.FlamingockStandalone;
 import io.flamingock.core.pipeline.Stage;
-import io.flamingock.examples.dynamodb.standalone.mongock.MongockLegacyDataProvisioner;
 import io.flamingock.oss.driver.dynamodb.driver.DynamoDBDriver;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 
+import java.net.URI;
 import java.net.URISyntaxException;
 
 public class CommunityStandaloneDynamoDBApp {
 
     public static void main(String[] args) throws URISyntaxException {
-        new CommunityStandaloneDynamoDBApp().run(DynamoDBUtil.getClient());
+        //DynamoDB Client Builder
+        DynamoDbClient client = DynamoDbClient.builder()
+                .region(Region.EU_WEST_1) // Set your AWS region
+                .endpointOverride(new URI("http://localhost:8000")) // Set your DynamoDB endpoint
+                .credentialsProvider(
+                        StaticCredentialsProvider.create(
+                                AwsBasicCredentials.create("dummye", "dummye") // Set your AWS credentials
+                        )
+                )
+                .build();
+
+        new CommunityStandaloneDynamoDBApp().run(client);
     }
 
     public void run(DynamoDbClient client) {
-        //This line adds data to simulate previous legacy Mongock executions
-        MongockLegacyDataProvisioner.provisionLegacyMongockData();
-
         //Running flamingock
         FlamingockStandalone.local()
                 .setDriver(new DynamoDBDriver(client))
-                .setMongockImporterConfiguration(CoreConfiguration.MongockImporterConfiguration.withSource("mongockChangeLog"))
                 .addStage(new Stage("stage-name")
-                        //adding mongock legacy changeUnits
-                        .addCodePackage("io.flamingock.examples.dynamodb.standalone.mongock")
-                        //Adding a package for new changeUnits (note that legacy and new changeUnits can coexist in the same package)
                         .addCodePackage("io.flamingock.examples.dynamodb.standalone.changes"))
                 .addDependency(client)
                 .setTransactionEnabled(true)
