+++ dynamodb/dynamodb-standalone/src/main/java/io/flamingock/examples/dynamodb/standalone/DynamoDBUtil.java
@@ -16,33 +16,15 @@
 
 package io.flamingock.examples.dynamodb.standalone;
 
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 import software.amazon.awssdk.services.dynamodb.model.*;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 public final class DynamoDBUtil {
     public DynamoDBUtil() {
-    }
-
-    public static DynamoDbClient getClient() throws URISyntaxException {
-        return DynamoDbClient.builder()
-                .region(Region.EU_WEST_1)
-                .endpointOverride(new URI("http://localhost:8000"))
-                .credentialsProvider(
-                        StaticCredentialsProvider.create(
-                                AwsBasicCredentials.create("dummye", "dummye")
-                        )
-                )
-                //.httpClient(UrlConnectionHttpClient.builder().build())
-                .build();
     }
 
     public static List<AttributeDefinition> getAttributeDefinitions(String pkName, String skName, String... vargs) {
@@ -145,7 +127,4 @@ public final class DynamoDBUtil {
             // Table already exists, continue
         }
     }
-
-
-
 }
