+++ dynamodb/dynamodb-standalone/src/main/java/io/flamingock/examples/dynamodb/standalone/changes/_3_insertAnotherUser_changeUnit.java
@@ -22,23 +22,19 @@ import io.flamingock.examples.dynamodb.standalone.UserEntity;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.model.PutItemEnhancedRequest;
+import software.amazon.awssdk.enhanced.dynamodb.model.TransactWriteItemsEnhancedRequest;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 
-@ChangeUnit(id = "insert-user", order = "3", transactional = false)
-public class _3_insertUser_changeUnit {
+@ChangeUnit(id = "insert-another-user", order = "4")
+public class _3_insertAnotherUser_changeUnit {
 
     @Execution
-    public void execution(DynamoDbClient client) {
+    public void execution(DynamoDbClient client, TransactWriteItemsEnhancedRequest.Builder writeRequestBuilder) {
         DynamoDbTable<UserEntity> table = DynamoDbEnhancedClient.builder()
                 .dynamoDbClient(client)
                 .build()
                 .table(UserEntity.tableName, TableSchema.fromBean(UserEntity.class));
 
-        table.putItem(
-                PutItemEnhancedRequest.builder(UserEntity.class)
-                        .item(new UserEntity("Pepe", "Pérez"))
-                        .build()
-        );
+        writeRequestBuilder.addPutItem(table, new UserEntity("Pablo", "López"));
     }
 }
